+++ activiti7-java/src/main/java/com/activiti/modules/controller/SysFormController.java
@@ -4,8 +4,10 @@ import java.util.Date;
 import java.util.List;
 
 import com.activiti.modules.entity.SysFormEntity;
+import com.activiti.modules.entity.SysUserEntity;
 import com.activiti.modules.service.SysFormService;
 import com.activiti.utils.R;
+import com.activiti.utils.exception.AException;
 import com.activiti.utils.page.PageUtils;
 import com.activiti.utils.page.TableDataInfo;
 import org.apache.commons.lang3.StringUtils;
@@ -57,6 +59,12 @@ public class SysFormController {
      */
     @PostMapping("save")
     public R<String> info(@RequestBody SysFormEntity model) {
+        if (StringUtils.isNotEmpty(model.getFormId())) {
+            SysFormEntity form = formService.getById(model.getFormId());
+            if (form == null) throw new AException("未知数据");
+            if (form.getIsSys() == 1) throw new AException("系统内置不能修改!");
+        }
+
         Date date = new Date();
         if (StringUtils.isEmpty(model.getFormId())) {
             model.setCreateTime(date);
@@ -73,6 +81,11 @@ public class SysFormController {
      */
     @DeleteMapping("delete")
     public R<String> delete(@RequestBody String id) {
+
+        SysFormEntity form = formService.getById(id);
+        if(form == null) throw new AException("未知数据");
+        if(form.getIsSys() == 1) throw new AException("系统内置不能删除!");
+
         formService.removeById(id);
         return R.ok();
     }
